requirement.txt
vosk
python-dotenv
os
PyAudio
transformers